# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chore Club API",
    "description": "Multi-user, multi-household, multi-device chore tracker. RFC 7807 errors, `If-Match` optimistic concurrency on aggregate roots, `Idempotency-Key` on side-effecting hot-path POSTs, and a delta-sync protocol on `/sync`. Authoritative spec: `docs/api-spec.md`.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local docker-compose stack."
    },
    {
      "url": "https://api.choreclub.app",
      "description": "Production."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 access token issued by `/v1/auth/google` or `/v1/auth/apple`. 15-min TTL."
      }
    },
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "description": "UUIDv7 string."
      },
      "IsoTimestamp": {
        "type": "string",
        "format": "date-time",
        "description": "ISO-8601 UTC, ms precision."
      },
      "RowVersion": {
        "type": "integer",
        "minimum": 0,
        "description": "Optimistic concurrency token (BIGINT UNSIGNED)."
      },
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 problem+json body. Returned for every error.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `household.not_a_member`."
          },
          "errors": {
            "description": "Optional structured payload (validation issues, conflict snapshot, etc.)."
          }
        }
      },
...
      "HouseholdRole": {
        "type": "string",
        "enum": [
          "adult",
          "child"
        ]
      },
      "HouseholdPermission": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member"
        ]
      },
...
      "Household": {
        "type": "object",
        "required": [
          "id",
          "name",
          "ownerUserId",
          "rowVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "ownerUserId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "rowVersion": {
            "$ref": "#/components/schemas/RowVersion"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          }
        }
      },
      "HouseholdMember": {
        "type": "object",
        "required": [
          "id",
          "householdId",
          "userId",
          "role",
          "permission",
          "displayName",
          "avatar",
          "color",
          "joinedAt",
          "removedAt",
          "rowVersion"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "householdId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "userId": {
            "$ref": "#/components/schemas/Uuid",
            "nullable": true,
            "description": "Null for standalone (non-account-backed) members."
          },
          "role": {
            "$ref": "#/components/schemas/HouseholdRole"
          },
          "permission": {
            "$ref": "#/components/schemas/HouseholdPermission"
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "avatar": {
            "type": "string",
            "maxLength": 16
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}([0-9a-fA-F]{2})?$"
          },
          "joinedAt": {
            "$ref": "#/components/schemas/IsoTimestamp"
          },
          "removedAt": {
            "$ref": "#/components/schemas/IsoTimestamp",
            "nullable": true
          },
          "rowVersion": {
            "$ref": "#/components/schemas/RowVersion"
          }
        }
      },
...
      "HouseholdCreateResponse": {
        "type": "object",
        "required": [
          "household",
          "member"
        ],
        "properties": {
          "household": {
            "$ref": "#/components/schemas/Household"
          },
          "member": {
            "$ref": "#/components/schemas/HouseholdMember"
          }
        }
      },
...
      "CreateHouseholdBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "RenameHouseholdBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "additionalProperties": false
      },
      "TransferOwnershipBody": {
        "type": "object",
        "properties": {
          "toMemberId": {
            "type": "string"
          }
        },
        "required": [
          "toMemberId"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/households": {
      "post": {
        "summary": "Create a household. Caller becomes owner.",
        "tags": [
          "households"
        ],
        "responses": {
          "201": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdBody"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "postHouseholds"
      }
    },
    "/v1/households/{householdId}": {
      "get": {
        "summary": "Get a household.",
        "tags": [
          "households"
        ],
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Quoted `rowVersion`."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Household"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — auth ok, permission insufficient.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-household-permission": "member",
        "operationId": "getHouseholdsByHouseholdId"
      },
      "patch": {
        "summary": "Rename a household. `If-Match` recommended.",
        "tags": [
          "households"
        ],
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Optimistic concurrency token. Quoted `rowVersion`. Mismatch → 409 `concurrency.row_version_stale`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Quoted `rowVersion`."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Household"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — auth ok, permission insufficient.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — stale `If-Match`, idempotency-key reuse, or domain conflict.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameHouseholdBody"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-household-permission": "admin",
        "operationId": "patchHouseholdsByHouseholdId"
      },
      "delete": {
        "summary": "Soft-delete a household and cascade to invites + tasks.",
        "tags": [
          "households"
        ],
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — auth ok, permission insufficient.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-household-permission": "owner",
        "operationId": "deleteHouseholdsByHouseholdId"
      }
    },
    "/v1/households/{householdId}/transfer-ownership": {
      "post": {
        "summary": "Transfer ownership to another account-backed member. Demotes prior owner to admin.",
        "tags": [
          "households"
        ],
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Household"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — auth ok, permission insufficient.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipBody"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-household-permission": "owner",
        "operationId": "postHouseholdsByHouseholdIdTransferOwnership"
      }
    },
...
  }
}